+++ src/components/generator/Generator.jsx
@@ -1,5 +1,6 @@
-import styled from "styled-components";
 
+
+import styled from "styled-components";
 import { useState } from "react";
 import DragDrop from "../drag-drop/DragDrop";
 import SpaceWorker from "../space-worker/SpaceWorker";
